wip tableをextensionに切り分け
